verificar entrada

+++ src/models/entries/deleteEntryModel.js
@@ -1,9 +1,9 @@
 import getPool from '../../db/getPool.js';
-const deleteEntryModel = async (entryId) => {
+const deleteEntryModel = async (entryId, userId) => {
     const pool = await getPool();
 
-    await pool.query(`DELETE FROM entries WHERE id = ?`,
-        [entryId]);
+    await pool.query(`DELETE FROM entries WHERE id = ? AND userId = ?`,
+        [entryId, userId]);
 
 };
